Integrate CertLogic library.
