+++ src/xcode/ENA/HealthCertificateToolkit/Tests/HealthCertificateToolkitTests/Validation/ValidationRulesTests.swift
@@ -23,4 +23,3 @@ class ValidationRulesTests: XCTestCase {
         XCTAssertEqual(validationRules.rules, rules)
     }
 }
-
